+++ src/main/java/bridge/BridgeMaker.java
@@ -3,8 +3,8 @@ package bridge;
 import java.util.ArrayList;
 import java.util.List;
 
-import static bridge.Direction.DOWN;
-import static bridge.Direction.UP;
+import static bridge.model.Direction.DOWN;
+import static bridge.model.Direction.UP;
 
 /**
  * 다리의 길이를 입력 받아서 다리를 생성해주는 역할을 한다.
