+++ src/main/java/bridge/dto/MapDTO.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.dto;
 
 import java.util.List;
