+++ src/main/java/bridge/message/NoticeMessage.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.message;
 
 public class NoticeMessage {
     public static final String START = "다리 건너기 게임을 시작합니다.";
