+++ src/main/java/bridge/model/Command.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.model;
 
 public enum Command {
     RETRY("R"),
