+++ src/main/java/bridge/model/Direction.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.model;
 
 public enum Direction {
     DOWN("D", 0),
