+++ src/main/java/bridge/view/InputView.java
@@ -1,8 +1,8 @@
-package bridge;
+package bridge.view;
 
 import camp.nextstep.edu.missionutils.Console;
 
-import static bridge.NoticeMessage.*;
+import static bridge.message.NoticeMessage.*;
 
 /**
  * 사용자로부터 입력을 받는 역할을 한다.
