+++ src/test/java/bridge/BridgeMakerTest.java
@@ -1,12 +1,11 @@
 package bridge;
 
+import bridge.model.Direction;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
-import static bridge.Direction.DOWN;
-import static bridge.Direction.UP;
 import static org.junit.jupiter.api.Assertions.*;
 
 class BridgeMakerTest {
